Fixed bug: The function did not read a value for the mesh variable.
It's sort of silly, because the 'mesh variable' is only a container
for attributes, and its value is never used. However, DAP requires a
value in order to send the result back to the client. The function
as it is now, works correctly with a file that conforms to 'ugrid 0.9'
although our code is somewhat lenient (e.g., while the 'cf_role'
attribute is required by the specification, our code will sort it out
if the data file otehrwise conforms to the spec but does not
include that attribute.